refactor: move ecstore rebalance worker helpers
